refactor(cli): move load-config.ts to src root, delete empty config/ folder
