+++ packages/cli/test/fixtures/honeycrisp-basic/smoke-test.ts
@@ -12,7 +12,7 @@
 import { join } from 'node:path';
 import { createWorkspace, dateTimeStringNow } from '@epicenter/workspace';
 import { filesystemPersistence } from '@epicenter/workspace/extensions/sync/desktop';
-import { loadConfig } from '../../../src/config/load-config';
+import { loadConfig } from '../../../src/load-config';
 import { definition } from './epicenter.config';
 
 const FIXTURE_DIR = import.meta.dir;
